split out compatiblity interface
